requirements.txt
flask
Flask-WTF
bootstrap-flask
requests
python-dotenv